requirements.txt
python-telegram-bot==13.14
pillow